fix(rubysonar):Add test for rational type
Rational type only support the syntx with ‘r’ suffix.
e.g.
a=‘2r’  #a is rational
a=Rational(2) #a is ?
Since type conversion for arithmetic calculus haven’t been implemented
yet in both pysonar2 and rubysonar, this is almost the only support for
rational type.
Regenerated all the test results and passed.